move u256/u512 defs into types/uint.rs
