+++ trin-core/src/portalnet/storage.rs
@@ -1,4 +1,4 @@
-use super::U256;
+use super::types::uint::U256;
 use crate::utils::{db::get_data_dir, distance::xor_two_values};
 use discv5::enr::NodeId;
 use hex;
