+++ trin-core/src/portalnet/types/mod.rs
@@ -1,2 +1,3 @@
 pub mod content_keys;
 pub mod messages;
+pub mod uint;
